requirements.txt
allure-pytest
jsonschema
requests
pytest
pytest-xdist
jsonref